Refactor naming packages
