--- src/main/java/com/book_app/dto/BookDto.java
@@ -1,15 +1,0 @@
-package com.book_app.dto;
-
-import java.math.BigDecimal;
-import lombok.Data;
-
-@Data
-public class BookDto {
-    private Long id;
-    private String title;
-    private String author;
-    private String isbn;
-    private BigDecimal price;
-    private String description;
-    private String coverImage;
-}
