--- src/main/java/com/book_app/dto/BookSearchParametersDto.java
@@ -1,8 +1,0 @@
-package com.book_app.dto;
-
-public record BookSearchParametersDto(String[] titles,
-                                      String[] authors,
-                                      String[] isbn,
-                                      String[] price,
-                                      String[] description) {
-}
