--- src/main/java/com/book_app/exception/EntityNotFoundException.java
@@ -1,7 +1,0 @@
-package com.book_app.exception;
-
-public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(String message) {
-        super(message);
-    }
-}
