--- src/main/java/com/book_app/repository/SpecificationBuilder.java
@@ -1,8 +1,0 @@
-package com.book_app.repository;
-
-import com.book_app.dto.BookSearchParametersDto;
-import org.springframework.data.jpa.domain.Specification;
-
-public interface SpecificationBuilder<T> {
-    Specification<T> build(BookSearchParametersDto bookSearchParametersDto);
-}
