--- src/main/java/com/book_app/repository/SpecificationProviderManager.java
@@ -1,5 +1,0 @@
-package com.book_app.repository;
-
-public interface SpecificationProviderManager<T> {
-    SpecificationProvider<T> getSpecificationProvider(String key);
-}
