+++ src/main/java/com/bookappstore/dto/BookDto.java
@@ -1,0 +1,15 @@
+package com.bookappstore.dto;
+
+import java.math.BigDecimal;
+import lombok.Data;
+
+@Data
+public class BookDto {
+    private Long id;
+    private String title;
+    private String author;
+    private String isbn;
+    private BigDecimal price;
+    private String description;
+    private String coverImage;
+}
