+++ src/main/java/com/bookappstore/dto/BookSearchParametersDto.java
@@ -1,0 +1,8 @@
+package com.bookappstore.dto;
+
+public record BookSearchParametersDto(String[] titles,
+                                      String[] authors,
+                                      String[] isbn,
+                                      String[] price,
+                                      String[] description) {
+}
