+++ src/main/java/com/bookappstore/exception/EntityNotFoundException.java
@@ -1,0 +1,7 @@
+package com.bookappstore.exception;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
