+++ src/main/java/com/bookappstore/exception/SpecificationProviderException.java
@@ -1,0 +1,7 @@
+package com.bookappstore.exception;
+
+public class SpecificationProviderException extends RuntimeException {
+    public SpecificationProviderException(String message) {
+        super(message);
+    }
+}
