+++ src/main/java/com/bookappstore/mapper/BookMapper.java
@@ -1,0 +1,16 @@
+package com.bookappstore.mapper;
+
+import com.bookappstore.config.MapperConfig;
+import com.bookappstore.dto.BookDto;
+import com.bookappstore.dto.CreateBookRequestDto;
+import com.bookappstore.model.Book;
+import org.mapstruct.Mapper;
+
+@Mapper(config = MapperConfig.class)
+public interface BookMapper {
+
+    BookDto toDto(Book book);
+
+    Book toModel(CreateBookRequestDto requestDto);
+
+}
