+++ src/main/java/com/bookappstore/repository/SpecificationBuilder.java
@@ -1,0 +1,8 @@
+package com.bookappstore.repository;
+
+import com.bookappstore.dto.BookSearchParametersDto;
+import org.springframework.data.jpa.domain.Specification;
+
+public interface SpecificationBuilder<T> {
+    Specification<T> build(BookSearchParametersDto bookSearchParametersDto);
+}
