+++ src/main/java/com/bookappstore/repository/SpecificationProvider.java
@@ -1,0 +1,9 @@
+package com.bookappstore.repository;
+
+import org.springframework.data.jpa.domain.Specification;
+
+public interface SpecificationProvider<T> {
+    String getKey();
+
+    Specification<T> getSpecification(String[] params);
+}
