+++ src/main/java/com/bookappstore/repository/SpecificationProviderManager.java
@@ -1,0 +1,5 @@
+package com.bookappstore.repository;
+
+public interface SpecificationProviderManager<T> {
+    SpecificationProvider<T> getSpecificationProvider(String key);
+}
