+++ src/main/java/com/bookappstore/repository/book/BookRepository.java
@@ -1,0 +1,8 @@
+package com.bookappstore.repository.book;
+
+import com.bookappstore.model.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+public interface BookRepository extends JpaRepository<Book, Long>, JpaSpecificationExecutor<Book> {
+}
