+++ src/main/java/com/bookappstore/repository/book/spec/PriceSpecificationProvider.java
@@ -1,0 +1,26 @@
+package com.bookappstore.repository.book.spec;
+
+import jakarta.persistence.criteria.Predicate;
+import com.bookappstore.model.Book;
+import com.bookappstore.repository.SpecificationProvider;
+import org.springframework.data.jpa.domain.Specification;
+import org.springframework.stereotype.Component;
+
+@Component
+public class PriceSpecificationProvider implements SpecificationProvider<Book> {
+    private static final String KEY_WORD = "price";
+
+    @Override
+    public String getKey() {
+        return KEY_WORD;
+    }
+
+    @Override
+    public Specification<Book> getSpecification(String[] params) {
+        return (root, query, criteriaBuilder) -> {
+            Predicate predicate = criteriaBuilder
+                    .lessThanOrEqualTo(root.get(KEY_WORD), params[0]);
+            return predicate;
+        };
+    }
+}
