+++ src/test/java/com/bookappstore/DemoApplicationTests.java
@@ -1,0 +1,14 @@
+package com.bookappstore;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
+
